[Feat/#78] VowelDrawingView를 구현하였습니다
